+++ src/main/java/es/uex/challengeapp/model/RetoDTO.java
@@ -1,19 +1,19 @@
 package es.uex.challengeapp.model;
 
 public class RetoDTO {
-    private Integer id;
+    private Long id;
     private String nombre;
     private String descripcion;
 
     // Constructor
-    public RetoDTO(Integer id, String nombre, String descripcion) {
+    public RetoDTO(Long id, String nombre, String descripcion) {
         this.id = id;
         this.nombre = nombre;
         this.descripcion = descripcion;
     }
 
     // Getters y setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
