+++ src/main/java/es/uex/challengeapp/repository/SubtareaRepository.java
@@ -10,6 +10,6 @@ import es.uex.challengeapp.model.Subtarea;
 @Repository
 public interface SubtareaRepository extends JpaRepository<Subtarea, Integer> {
 
-	List<Subtarea> findByRetoComplejoId(Integer retoComplejoId);
+	List<Subtarea> findByRetoComplejoId(Long retoComplejoId);
 
 }
